+++ routes/api/email-subscribe.js
@@ -9,7 +9,7 @@ const error = {
         code: 100
     },
     101: {
-        message: 'Email not provided',
+        message: 'Email / id not provided',
         code: 101
     },
     102: {
@@ -20,6 +20,10 @@ const error = {
         message: 'Unknown Error',
         code: 103
     },
+    104: {
+        message: 'Email is not in the list',
+        code: 103
+    },
 }
 
 const status = {
@@ -28,8 +32,16 @@ const status = {
         code: 200,
     },
     201: {
-        message: "Cant add to mailing list",
+        message: "Removed Email from mailing list",
         code: 201,
+    },
+    202: {
+        message: "Cant add to mailing list",
+        code: 202,
+    },
+    203: {
+        message: "Cant Remove to mailing list",
+        code: 203,
     },
 }
 router.post('/', async (req, res) => {    
@@ -38,14 +50,14 @@ router.post('/', async (req, res) => {
     if (!email) {
         return res.status(400).json({
             error: error[101],
-            status: status[201]
+            status: status[202]
         })
     }
     const check = await emailList.findOne({ email: email });
     if(check) {
         return res.status(401).json({
             error: error[102],
-            status: status[201]
+            status: status[202]
         })
     }
 
@@ -66,9 +78,41 @@ router.post('/', async (req, res) => {
         console.log('subscribe api - ' + err)
         return res.status(400).json({
             error: error[103],
-            status: status[201]
+            status: status[202]
         })
     }
 })
 
+router.delete('/', async (req, res) => {
+    const { id } = req.body;
+    if (!id) {
+        return res.status(400).json({
+            error: error[101],
+            status: status[203]
+        })
+    }
+    const check = await emailList.findOne({ _id: id });
+    if(!check) {
+        return res.status(400).json({
+            error: error[104],
+            status: status[203]
+        })
+    }
+
+    try {
+        await emailList.deleteOne({ _id: id })
+        res.status(200).json({
+            error: error[100],
+            status: status[201]
+        })
+        return;
+    } catch (err) {
+        console.log('unsubscribe api (routes/api/email-subscribe.js) - ' + err)
+        return res.status(400).json({
+            error: error[103],
+            status: status[202]
+        })
+    }
+
+})
 module.exports = router;